Move/Refactor FailedFuture and add comments
